+++ app/src/test/java/com/techcourse/service/AppUserServiceTest.java
@@ -7,14 +7,11 @@ import com.techcourse.domain.User;
 import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class UserServiceTest {
-
+class AppUserServiceTest {
     private JdbcTemplate jdbcTemplate;
     private UserDao userDao;
 
@@ -31,7 +28,7 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, DataSourceConfig.getInstance());
+        final var userService = new AppUserService(userDao, userHistoryDao);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -40,21 +37,5 @@ class UserServiceTest {
         final var actual = userService.findById(1L);
 
         assertThat(actual.getPassword()).isEqualTo(newPassword);
-    }
-
-    @Test
-    void testTransactionRollback() {
-        final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(userDao, userHistoryDao, DataSourceConfig.getInstance());
-
-        final var newPassword = "newPassword";
-        final var createBy = "gugu";
-
-        assertThrows(DataAccessException.class,
-                () -> userService.changePassword(1L, newPassword, createBy));
-
-        final var actual = userService.findById(1L);
-
-        assertThat(actual.getPassword()).isNotEqualTo(newPassword);
     }
 }
