+++ app/src/test/java/com/techcourse/service/TxUserServiceTest.java
@@ -1,0 +1,45 @@
+package com.techcourse.service;
+
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.dao.UserDao;
+import com.techcourse.domain.User;
+import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class TxUserServiceTest {
+    private JdbcTemplate jdbcTemplate;
+    private UserDao userDao;
+
+    @BeforeEach
+    void setUp() {
+        this.jdbcTemplate = new JdbcTemplate(DataSourceConfig.getInstance());
+        this.userDao = new UserDao(jdbcTemplate);
+
+        DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
+        final var user = new User("gugu", "password", "hkkang@woowahan.com");
+        userDao.insert(user);
+    }
+
+    @Test
+    void testTransactionRollback() {
+        final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(appUserService);
+
+        final var newPassword = "newPassword";
+        final var createBy = "gugu";
+        
+        assertThrows(DataAccessException.class,
+                () -> userService.changePassword(1L, newPassword, createBy));
+
+        final var actual = userService.findById(1L);
+
+        assertThat(actual.getPassword()).isNotEqualTo(newPassword);
+    }
+}
